# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Car Wash - Company (Empresa) API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResult {
  message: string;
}

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  pending: 'pending',
  accepted: 'accepted',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface CatalogRef {
  id: string;
  slug: string;
  name: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  client: 'client',
  company: 'company',
  admin: 'admin',
} as const;

export interface User {
  id: string;
  name: string;
  email: string;
  phone: string;
  role: UserRole;
  /** @nullable */
  companyId?: string | null;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthResult {
  user: User;
  token: string;
}

export interface CompanyProfile {
  id: string;
  name: string;
  email: string;
  phone: string;
  active: boolean;
  /** @nullable */
  rating?: number | null;
  totalBookings: number;
  completedBookings: number;
  createdAt: string;
}

export interface CompanyProfileUpdate {
  name?: string;
  phone?: string;
}

export type CompanyBookingCompanyStatus = typeof CompanyBookingCompanyStatus[keyof typeof CompanyBookingCompanyStatus];


export const CompanyBookingCompanyStatus = {
  pending_acceptance: 'pending_acceptance',
  accepted_by_company: 'accepted_by_company',
  rejected_by_company: 'rejected_by_company',
} as const;

export interface CompanyBooking {
  id: string;
  clientName: string;
  clientPhone: string;
  addressFull: string;
  vehicleSize: CatalogRef;
  /** @nullable */
  vehicleBrand?: string | null;
  /** @nullable */
  vehicleModel?: string | null;
  /** @nullable */
  vehicleColor?: string | null;
  /** @nullable */
  vehiclePlate?: string | null;
  washType: CatalogRef;
  addOns: CatalogRef[];
  date: string;
  time: string;
  totalPrice: number;
  status: BookingStatus;
  companyStatus: CompanyBookingCompanyStatus;
  /** @nullable */
  comments?: string | null;
  createdAt: string;
}

export interface Pagination {
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface CompanyBookingList {
  data: CompanyBooking[];
  pagination: Pagination;
}

export interface RejectInput {
  reason: string;
}

export type CompanyDashboardSummary = {
  total: number;
  pendingAcceptance: number;
  accepted: number;
  inProgress: number;
  completed: number;
  cancelled: number;
  revenueToday: number;
};

export interface CompanyDashboard {
  date: string;
  summary: CompanyDashboardSummary;
  nextBooking: CompanyBooking | null;
  upcoming: CompanyBooking[];
}

export interface CompanySlot {
  time: string;
  enabled: boolean;
}

export interface CompanyAvailability {
  slots: CompanySlot[];
  blockedDates: string[];
}

export interface CompanyAvailabilityUpdate {
  slots?: CompanySlot[];
  blockedDates?: string[];
}

export interface CompanyPackageOption {
  package: CatalogRef;
  name: string;
  color: string;
  washType: CatalogRef;
  active: boolean;
}

export interface CompanyPackageToggle {
  packageId: string;
  active: boolean;
}

export interface CompanyPackagesUpdate {
  packages: CompanyPackageToggle[];
}

export type CompanyServiceOptionKind = typeof CompanyServiceOptionKind[keyof typeof CompanyServiceOptionKind];


export const CompanyServiceOptionKind = {
  wash_type: 'wash_type',
  add_on: 'add_on',
} as const;

export interface CompanyServiceOption {
  service: CatalogRef;
  name: string;
  price: number;
  active: boolean;
  kind: CompanyServiceOptionKind;
}

export interface CompanyServiceToggle {
  serviceId: string;
  active: boolean;
}

export interface CompanyServicesUpdate {
  services: CompanyServiceToggle[];
}

export type CompanyEarningsByType = {
  directo: number;
  membresia: number;
};

export type CompanyEarningsDailySeriesItem = {
  date: string;
  amount: number;
  services: number;
};

export interface CompanyEarnings {
  dateFrom: string;
  dateTo: string;
  totalServices: number;
  totalAmount: number;
  pending: number;
  paid: number;
  byType: CompanyEarningsByType;
  dailySeries: CompanyEarningsDailySeriesItem[];
}

export type CompanyBillingPaymentType = typeof CompanyBillingPaymentType[keyof typeof CompanyBillingPaymentType];


export const CompanyBillingPaymentType = {
  directo: 'directo',
  membresia: 'membresia',
} as const;

export type CompanyBillingPaymentStatus = typeof CompanyBillingPaymentStatus[keyof typeof CompanyBillingPaymentStatus];


export const CompanyBillingPaymentStatus = {
  pendiente: 'pendiente',
  pagado: 'pagado',
} as const;

export interface CompanyBilling {
  id: string;
  bookingId: string;
  date: string;
  time: string;
  clientName: string;
  vehicleSize: CatalogRef;
  washType: CatalogRef;
  addOns: CatalogRef[];
  amount: number;
  paymentType: CompanyBillingPaymentType;
  paymentStatus: CompanyBillingPaymentStatus;
  /** @nullable */
  paidAt?: string | null;
}

export interface CompanyBillingList {
  data: CompanyBilling[];
  pagination: Pagination;
}

export type GetCompanyDashboardParams = {
/**
 * YYYY-MM-DD (default today)
 */
date?: string;
};

export type ListCompanyBookingsParams = {
date?: string;
dateFrom?: string;
dateTo?: string;
companyStatus?: ListCompanyBookingsCompanyStatus;
status?: BookingStatus;
search?: string;
page?: number;
limit?: number;
};

export type ListCompanyBookingsCompanyStatus = typeof ListCompanyBookingsCompanyStatus[keyof typeof ListCompanyBookingsCompanyStatus];


export const ListCompanyBookingsCompanyStatus = {
  pending_acceptance: 'pending_acceptance',
  accepted_by_company: 'accepted_by_company',
  rejected_by_company: 'rejected_by_company',
} as const;

export type GetCompanyEarningsParams = {
dateFrom?: string;
dateTo?: string;
};

export type ListCompanyEarningServicesParams = {
dateFrom?: string;
dateTo?: string;
paymentStatus?: ListCompanyEarningServicesPaymentStatus;
paymentType?: ListCompanyEarningServicesPaymentType;
page?: number;
limit?: number;
};

export type ListCompanyEarningServicesPaymentStatus = typeof ListCompanyEarningServicesPaymentStatus[keyof typeof ListCompanyEarningServicesPaymentStatus];


export const ListCompanyEarningServicesPaymentStatus = {
  pendiente: 'pendiente',
  pagado: 'pagado',
} as const;

export type ListCompanyEarningServicesPaymentType = typeof ListCompanyEarningServicesPaymentType[keyof typeof ListCompanyEarningServicesPaymentType];


export const ListCompanyEarningServicesPaymentType = {
  directo: 'directo',
  membresia: 'membresia',
} as const;