# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Car Wash - Company (Empresa) API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
...
  /company/dashboard:
    get:
      operationId: getCompanyDashboard
      tags: [bookings]
      summary: Wow endpoint - dashboard summary for today
      parameters:
        - in: query
          name: date
          schema: { type: string }
          description: "YYYY-MM-DD (default today)"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyDashboard"
  /company/bookings:
    get:
      operationId: listCompanyBookings
      tags: [bookings]
      summary: List bookings assigned to this company
      parameters:
        - in: query
          name: date
          schema: { type: string }
        - in: query
          name: dateFrom
          schema: { type: string }
        - in: query
          name: dateTo
          schema: { type: string }
        - in: query
          name: companyStatus
          schema:
            type: string
            enum: [pending_acceptance, accepted_by_company, rejected_by_company]
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/BookingStatus"
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyBookingList"
  /company/bookings/{bookingId}:
    get:
      operationId: getCompanyBooking
      tags: [bookings]
      parameters:
        - in: path
          name: bookingId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyBooking"
  /company/bookings/{bookingId}/accept:
    put:
      operationId: acceptCompanyBooking
      tags: [bookings]
      parameters:
        - in: path
          name: bookingId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyBooking"
  /company/bookings/{bookingId}/reject:
    put:
      operationId: rejectCompanyBooking
      tags: [bookings]
      parameters:
        - in: path
          name: bookingId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyBooking"
  /company/bookings/{bookingId}/start:
    put:
      operationId: startCompanyBooking
      tags: [bookings]
      parameters:
        - in: path
          name: bookingId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyBooking"
  /company/bookings/{bookingId}/complete:
    put:
      operationId: completeCompanyBooking
      tags: [bookings]
      parameters:
        - in: path
          name: bookingId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyBooking"
...
components:
  schemas:
...
    BookingStatus:
      type: string
      enum: [pending, accepted, in_progress, completed, cancelled]
    CatalogRef:
      type: object
      required: [id, slug, name]
      properties:
        id: { type: string }
        slug: { type: string }
        name: { type: string }
    User:
      type: object
      required: [id, name, email, phone, role]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        role:
          type: string
          enum: [client, company, admin]
        companyId: { type: ["string", "null"] }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    AuthResult:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token: { type: string }
...
    CompanyBooking:
      type: object
      required:
        - id
        - clientName
        - clientPhone
        - addressFull
        - vehicleSize
        - washType
        - addOns
        - date
        - time
        - totalPrice
        - status
        - companyStatus
        - createdAt
      properties:
        id: { type: string }
        clientName: { type: string }
        clientPhone: { type: string }
        addressFull: { type: string }
        vehicleSize:
          $ref: "#/components/schemas/CatalogRef"
        vehicleBrand: { type: ["string", "null"] }
        vehicleModel: { type: ["string", "null"] }
        vehicleColor: { type: ["string", "null"] }
        vehiclePlate: { type: ["string", "null"] }
        washType:
          $ref: "#/components/schemas/CatalogRef"
        addOns:
          type: array
          items:
            $ref: "#/components/schemas/CatalogRef"
        date: { type: string }
        time: { type: string }
        totalPrice: { type: number }
        status:
          $ref: "#/components/schemas/BookingStatus"
        companyStatus:
          type: string
          enum: [pending_acceptance, accepted_by_company, rejected_by_company]
        comments: { type: ["string", "null"] }
        createdAt: { type: string }
    Pagination:
      type: object
      required: [total, page, limit, totalPages]
      properties:
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        totalPages: { type: integer }
    CompanyBookingList:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/CompanyBooking"
        pagination:
          $ref: "#/components/schemas/Pagination"
    RejectInput:
      type: object
      required: [reason]
      properties:
        reason: { type: string }
    CompanyDashboard:
      type: object
      required: [date, summary, nextBooking, upcoming]
      properties:
        date: { type: string }
        summary:
          type: object
          required:
            - total
            - pendingAcceptance
            - accepted
            - inProgress
            - completed
            - cancelled
            - revenueToday
          properties:
            total: { type: integer }
            pendingAcceptance: { type: integer }
            accepted: { type: integer }
            inProgress: { type: integer }
            completed: { type: integer }
            cancelled: { type: integer }
            revenueToday: { type: number }
        nextBooking:
          anyOf:
            - $ref: "#/components/schemas/CompanyBooking"
            - type: "null"
        upcoming:
          type: array
          items:
            $ref: "#/components/schemas/CompanyBooking"
...