# lib/api-zod/src/generated/types/authResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Car Wash - Company (Empresa) API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface AuthResult {
  user: User;
  token: string;
}