# lib/api-zod/src/generated/types/bookingStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Car Wash - Company (Empresa) API
 * OpenAPI spec version: 0.1.0
 */

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  pending: 'pending',
  accepted: 'accepted',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;