# lib/api-zod/src/generated/types/catalogRef.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Car Wash - Company (Empresa) API
 * OpenAPI spec version: 0.1.0
 */

export interface CatalogRef {
  id: string;
  slug: string;
  name: string;
}